Automatically include all packages specified in a pipeline configuration
