+++ sherlock/microscope/__main__.py
@@ -5,10 +5,8 @@ python -m sherlock.microscope.server \
 ```
 """
 import argparse
-import importlib
 import logging
 import os
-import pkgutil
 import sys
 from pathlib import Path
 from typing import Optional
@@ -127,34 +125,6 @@ def make_app(
     return app
 
 
-def _import_submodules(package_name: str) -> None:
-    """
-    Import all submodules under the given package.
-    Primarily useful to have custom classes get loaded and registered.
-    """
-    # Code taken from https://github.com/allenai/allennlp/blob/v0.9.0/allennlp/common/util.py#L308
-    importlib.invalidate_caches()
-
-    # For some reason, python doesn't always add this by default to your path, but you pretty much
-    # always want it when using `--include-package`.  And if it's already there, adding it again at
-    # the end won't hurt anything.
-    sys.path.append(".")
-
-    # Import at top level
-    module = importlib.import_module(package_name)
-    path = getattr(module, "__path__", [])
-    path_string = "" if not path else path[0]
-
-    # walk_packages only finds immediate children, so need to recurse.
-    for module_finder, name, _ in pkgutil.walk_packages(path):
-        # Sometimes when you import third-party libraries that are on your path,
-        # `pkgutil.walk_packages` returns those too, so we need to skip them.
-        if path_string and module_finder.path != path_string:
-            continue
-        subpackage = f"{package_name}.{name}"
-        _import_submodules(subpackage)
-
-
 def main(args):
     parser = argparse.ArgumentParser(description="Serve up a simple model")
 
@@ -162,13 +132,6 @@ def main(args):
         "--pipeline-config", type=str, required=True, help="path to the pipeline configuration file"
     )
     parser.add_argument("--cuda-device", type=int, default=-1, help="id of GPU to use (if any)")
-    parser.add_argument(
-        "--include-package",
-        type=str,
-        action="append",
-        default=[],
-        help="additional packages to include",
-    )
     parser.add_argument(
         "-o",
         "--overrides",
@@ -189,10 +152,6 @@ def main(args):
     )
 
     args = parser.parse_args(args)
-
-    # Import any additional modules needed (to register custom classes)
-    for package_name in getattr(args, "include_package", ()):
-        _import_submodules(package_name)
 
     pipeline = Pipeline.from_file(args.pipeline_config, args.cuda_device)
